+++ MANIFEST.in
@@ -3,4 +3,5 @@ include README.md
 include requirements.txt
 recursive-include mthree/src *.pyx
 recursive-include mthree/src *.pxd
+recursive-include mthree/test/src *.pyx
 recursive-include mthree/test/data *.json
