+++ mthree/generators/random.py
@@ -18,7 +18,7 @@ from mthree.exceptions import M3Error
 
 class RandomGenerator:
     """Random bit-array generator"""
-    def __init__(self, num_qubits, num_arrays=None, seed=None):
+    def __init__(self, num_qubits, num_arrays=16, seed=None):
         """Generator of random arrays corresponding to random x-gates on
         qubits for TexMex mitigation
 
@@ -26,12 +26,12 @@ class RandomGenerator:
 
         Parameters:
             num_qubits (int): Number of qubits
-            num_arrays (int): Number of arrays to generate, default=None (infinite)
+            num_arrays (int): Number of arrays to generate, default=16 (None=infinite)
             seed (int): seed for RNG, default=None
 
         Attributes:
             num_qubits (int): Number of qubits / length of arrays
-            length (int): Total number of generated arrays, default=None (infinite)
+            length (int): Total number of generated arrays, default=16 (None=infinite)
             seed (int): Seed used for RNG
         """
         self.seed = seed
