+++ mthree/test/test_converters.py
@@ -13,7 +13,7 @@
 
 """Test the converters"""
 from mthree.matrix import bitstring_int
-from .converters_testing import _test_counts_roundtrip, _test_counts_to_array
+from .src.converters_testing import _test_counts_roundtrip, _test_counts_to_array
 
 COUNTS = {'00000': 520,
           '00001': 10,
